# public/openapi.yaml
openapi: 3.1.0
info:
  title: Cimbra API v1
  version: '2026-09-01'
  description: Versioned financial infrastructure API backed by a persistent double-entry ledger, scoped S2S API keys and signed durable webhooks. This deployment is a sandbox and does not move real funds. Every response includes X-Request-Id and Cimbra-Version; responses after successful API-key authorization also expose X-RateLimit-Limit, X-RateLimit-Remaining and X-RateLimit-Reset. Mutations replayed from an idempotency key include Idempotent-Replayed, while error responses may include Cimbra-Should-Retry so official SDKs can make safe retry decisions.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
...
servers:
  - url: https://cimbra-rose.vercel.app
    description: Persistent sandbox. Does not move real funds.
security: []
paths:
...
  /api/v1/rail-instruments:
    get:
      operationId: listRailInstruments
      summary: List sandbox CVU and tenant aliases
      description: Instruments issued by Cimbra as a sandbox PSP. Prefix 0009999 is not a Coelsa assignment. Aliases resolve only inside the tenant.
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200': { description: Rail instrument page }
        '403': { description: Missing transfers:read scope or console role }
    post:
      operationId: issueRailInstruments
      summary: Issue a sandbox CVU and optional alias for an ARS Argentine account
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [accountId]
              properties:
                accountId: { type: string, format: uuid }
                alias: { type: [string, 'null'], minLength: 6, maxLength: 20 }
      responses:
        '201': { description: CVU and optional alias issued }
        '200': { description: Idempotent replay }
        '400': { description: Invalid instrument payload }
        '403': { description: Missing transfers:write scope or finance.write permission }
        '409': { description: CVU already issued, alias conflict or idempotency mismatch }
...
  /api/v1/payment-qrs:
    get:
      operationId: listPaymentQrs
      summary: List Cimbra QR collection payloads
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200': { description: Payment QR page }
        '403': { description: Missing transfers:read scope or console role }
    post:
      operationId: createPaymentQr
      summary: Create a Cimbra-owned QR payload for an account
      description: Payload format cimbra:qr:v1 is not the Argentine interoperable QR. Payment is tenant-internal.
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [accountId, description]
              properties:
                accountId: { type: string, format: uuid }
                amount:
                  oneOf:
                    - { type: string, pattern: '^\d+(?:\.\d+)?$' }
                    - { type: number, exclusiveMinimum: 0, maximum: 10000000 }
                currency: { type: string, enum: [ARS] }
                description: { type: string, minLength: 2, maxLength: 180 }
                expiresInMinutes: { type: integer, minimum: 5, maximum: 1440 }
      responses:
        '201': { description: QR created }
        '200': { description: Idempotent replay }
        '400': { description: Invalid payload }
        '403': { description: Missing transfers:write scope or finance.write permission }
  /api/v1/payment-qrs/{id}/pay:
    post:
      operationId: payPaymentQr
      summary: Pay a Cimbra QR from another account in the tenant
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [sourceAccountId, externalReference]
              properties:
                sourceAccountId: { type: string, format: uuid }
                amount:
                  oneOf:
                    - { type: string, pattern: '^\d+(?:\.\d+)?$' }
                    - { type: number, exclusiveMinimum: 0, maximum: 10000000 }
                externalReference: { type: string, minLength: 2, maxLength: 100 }
                signals: { $ref: '#/components/schemas/RiskSignalsInput' }
      responses:
        '201': { description: QR paid }
        '200': { description: Idempotent replay }
        '400': { description: Invalid payload or open-amount QR without amount }
        '403': { description: Missing transfers:write scope or finance.write permission }
        '404': { description: QR not found }
        '409': { description: QR inactive or expired }
        '422': { description: Amount mismatch, insufficient funds or risk decline }
  /api/v1/payment-links:
    get:
      operationId: listPaymentLinks
      summary: List sandbox collection payment links
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200': { description: Payment link page }
        '403': { description: Missing payments:read scope or console role }
    post:
      operationId: createPaymentLink
      summary: Create a sandbox collection link against an Argentine ARS account
      description: Payload format cimbra:link:v1 is not a card network checkout. Allowed methods are internal Cimbra accounts or sandbox inbound cash-in. Card, POS and interoperable QR return 422.
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [accountId, externalReference, description, amount]
              properties:
                accountId: { type: string, format: uuid }
                externalReference: { type: string, minLength: 2, maxLength: 100 }
                description: { type: string, minLength: 2, maxLength: 180 }
                amount:
                  oneOf:
                    - { type: string, pattern: '^\d+(?:\.\d+)?$' }
                    - { type: number, exclusiveMinimum: 0, maximum: 10000000 }
                currency: { type: string, enum: [ARS] }
                expiresInMinutes: { type: integer, minimum: 5, maximum: 10080 }
                methods:
                  type: array
                  minItems: 1
                  items: { type: string, enum: [internal, sandbox_inbound] }
      responses:
        '201': { description: Payment link created }
        '200': { description: Idempotent replay }
        '400': { description: Invalid payload }
        '403': { description: Missing payments:write scope or finance.write permission }
        '422': { description: Card, POS or interoperable QR requested }
...
components:
  parameters:
    ResourceId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    PageLimit:
      in: query
      name: limit
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    PageCursor:
      in: query
      name: cursor
      required: false
      description: Opaque cursor returned as nextCursor by the previous page.
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Stable unique key for this logical write. Replays return the original resource; reusing it with a different payload returns 409.
      schema: { type: string, minLength: 8, maxLength: 100, pattern: '^[A-Za-z0-9._:-]+$' }
  schemas:
...
    RiskSignalsInput:
      type: object
      additionalProperties: false
      properties:
        deviceReference: { type: string, minLength: 2, maxLength: 160, writeOnly: true }
        identityReference: { type: string, minLength: 2, maxLength: 160, writeOnly: true }
        deviceTrust: { type: string, enum: [trusted, unknown, suspicious] }
        identityVerified: { type: boolean }
        ipCountry: { type: string, pattern: '^[A-Za-z]{2}$' }
        countryMismatch: { type: boolean }
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-cimbra_session
      description: First-party opaque session token. Only its SHA-256 hash is stored server-side.
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: Cimbra API key
      description: S2S key with a cim_sk_test_ prefix. The plaintext secret is shown once and only its SHA-256 hash is stored.